+++ src/main/java/lonestarrr/arconia/data/ArconiaBiomeTagsProvider.java
@@ -21,36 +21,20 @@ public class ArconiaBiomeTagsProvider extends BiomeTagsProvider {
     @Override
     protected void addTags(HolderLookup.@NotNull Provider lookupProvider) {
         tag(Tags.Biomes.HAS_CLOVER)
-                .add(Biomes.PLAINS)
-                .add(Biomes.SUNFLOWER_PLAINS)
-                .add(Biomes.FOREST)
-                .add(Biomes.FLOWER_FOREST)
-                .add(Biomes.BIRCH_FOREST)
-                .add(Biomes.OLD_GROWTH_BIRCH_FOREST)
-                .add(Biomes.DARK_FOREST)
-                .add(Biomes.CHERRY_GROVE)
-                .add(Biomes.MEADOW)
-                .add(Biomes.OLD_GROWTH_PINE_TAIGA)
-                .add(Biomes.OLD_GROWTH_SPRUCE_TAIGA)
-                .add(Biomes.TAIGA)
-                .add(Biomes.JUNGLE)
-                .add(Biomes.SPARSE_JUNGLE)
-                .add(Biomes.SAVANNA)
-                .add(Biomes.SAVANNA_PLATEAU)
-                .add(Biomes.WINDSWEPT_SAVANNA)
-                .add(Biomes.WINDSWEPT_FOREST)
-                .add(Biomes.WINDSWEPT_HILLS);
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_DECIDUOUS_TREE)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_CONIFEROUS_TREE)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_JUNGLE_TREE)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_FLORAL)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_PLAINS)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_HILL)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_OLD_GROWTH)
+        ;
         tag(Tags.Biomes.HAS_ARCONIUM_TREES)
-                .add(Biomes.FOREST)
-                .add(Biomes.FLOWER_FOREST)
-                .add(Biomes.BIRCH_FOREST)
-                .add(Biomes.OLD_GROWTH_BIRCH_FOREST)
-                .add(Biomes.OLD_GROWTH_PINE_TAIGA)
-                .add(Biomes.OLD_GROWTH_SPRUCE_TAIGA)
-                .add(Biomes.TAIGA)
-                .add(Biomes.JUNGLE)
-                .add(Biomes.SAVANNA)
-                .add(Biomes.WINDSWEPT_FOREST);
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_DECIDUOUS_TREE)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_CONIFEROUS_TREE)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_JUNGLE_TREE)
+                .addTag(net.neoforged.neoforge.common.Tags.Biomes.IS_OLD_GROWTH)
+                ;
     }
 
 }